Feature: test users endpoint with Jest - Should get all users from the database - Should create a new user - Should not create new suer without username - make use of supertest to make API requests from within the test suite.
