+++ package.json
@@ -4,7 +4,8 @@
   "description": "This is a demo rest api project",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon index.js"
+    "start": "nodemon index.js",
+    "test": "jest"
   },
   "keywords": [],
   "author": "Roland-Sankara",
@@ -17,5 +18,9 @@
     "jsonwebtoken": "^8.5.1",
     "nodemon": "^2.0.16",
     "prisma": "^3.15.2"
+  },
+  "devDependencies": {
+    "jest": "^28.1.3",
+    "supertest": "^6.2.4"
   }
 }
